Updated the JavaScript to work with multiple keys per reference
Moved several signatures to use SystemReference instead of Reference
(because of the need for ReferenceNames)
Added a clone method on SystemReference so I could change the name
without it persisting
Pulled the stitch.js into a file and embedded it as a resource so it's
easier to edit
Rewrote stitch.js to work with the key based system we have now. Removed
extra stuff we don't need from it.
Modified referencer to send through all keys as part of the url
Created a ToSystemReference that should convert any reference to a
system reference as long as you pass in a root path